Making more generic
Attempting to make this slightly more generic in nature so that it can
be used for different types of radars.

The radar.html is now generic, and assumes that you are using the
radarData.js file ( the expectation is that you will simply rename the
data file you want to now reference. The goals is to change the fewest
things possible ).

The title for the html page is now driven from the radarData.js

The radar arcs is now no longer hard coded to 4, but can be N , and the
size and contents is determined by the radar_arcs now in RadarData.js

The data is now able to determine the size of the blip to allow users to
use the size of the blip to  relay extra information.

Each Blip can now have an optional URL associated with it, and if it
does the cursor will turn into a pointer.

Removed unused code to clean up a bit.